* Update Ad mob credentials * Code refactor * Update gitignore file
