+++ .gitignore
@@ -12,3 +12,5 @@
 /captures
 .externalNativeBuild
 .cxx
+secret.properties
+lol.keystore
