+++ .idea/misc.xml
@@ -5,7 +5,7 @@
       <map>
         <entry key="app/src/main/res/drawable/ic_bowling.xml" value="0.1365" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.1203125" />
-        <entry key="app/src/main/res/layout/app_bar_main.xml" value="0.2578125" />
+        <entry key="app/src/main/res/layout/app_bar_main.xml" value="0.1" />
         <entry key="app/src/main/res/layout/christmas_fragment.xml" value="0.2578125" />
         <entry key="app/src/main/res/layout/content_main.xml" value="0.2578125" />
         <entry key="app/src/main/res/layout/fragment_about.xml" value="0.21203703703703702" />
