+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/AppTheme">
         <meta-data
             android:name="com.google.android.gms.ads.APPLICATION_ID"
-            android:value="ca-app-pub-1982812619462025~9638790873" />
+            android:value="@string/ad_value" />
 
         <activity
             android:name="com.project.segunfrancis.lol.ui.MainActivity"
