+++ app/src/main/java/com/project/segunfrancis/lol/repository/LolRepository.kt
@@ -1,6 +1,6 @@
 package com.project.segunfrancis.lol.repository
 
-import com.project.segunfrancis.lol.api.model.JokeResponse
+import com.project.segunfrancis.lol.data.model.JokeResponse
 
 interface LolRepository {
     suspend fun getJokes(category: String): JokeResponse
