+++ app/src/main/java/com/project/segunfrancis/lol/ui/any/AnyFragment.kt
@@ -59,7 +59,7 @@ class AnyFragment : Fragment(R.layout.fragment_any) {
     private fun handleError(error: Throwable) = with(binding) {
         genericInclude.shuffleButton.showMessage(
             error.localizedMessage,
-            R.string.text_retry,
+            R.string.retry,
             indefiniteDuration = true
         ) {
             viewModel.getAnyJoke(JokeCategory.ANY.value)
