+++ app/src/main/java/com/project/segunfrancis/lol/ui/miscellaneous/MiscellaneousFragment.kt
@@ -59,7 +59,7 @@ class MiscellaneousFragment : Fragment(R.layout.fragment_miscellaneous) {
     private fun handleError(error: Throwable) = with(binding) {
         genericInclude.shuffleButton.showMessage(
             error.localizedMessage,
-            R.string.text_retry,
+            R.string.retry,
             indefiniteDuration = true
         ) {
             viewModel.getMiscJoke(JokeCategory.MISC.value)
