+++ app/src/main/java/com/project/segunfrancis/lol/ui/pun/PunFragment.kt
@@ -59,7 +59,7 @@ class PunFragment : Fragment(R.layout.pun_fragment) {
     private fun handleError(error: Throwable) = with(binding) {
         genericInclude.shuffleButton.showMessage(
             error.localizedMessage,
-            R.string.text_retry,
+            R.string.retry,
             indefiniteDuration = true
         ) {
             viewModel.getPunJoke(JokeCategory.PUN.value)
