+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -7,44 +7,44 @@
         <item
             android:id="@+id/nav_any"
             android:icon="@drawable/ic_emoticon_lol"
-            android:title="@string/menu_any" />
+            android:title="@string/any_jokes" />
         <item
             android:id="@+id/nav_programming"
             android:icon="@drawable/ic_laptop"
-            android:title="@string/menu_programming" />
+            android:title="@string/programming_jokes" />
         <item
             android:id="@+id/nav_dark"
             android:icon="@drawable/ic_emoticon_devil"
-            android:title="@string/menu_dark" />
+            android:title="@string/dark_jokes" />
         <item
             android:id="@+id/nav_pun"
             android:icon="@drawable/ic_pun"
-            android:title="@string/menu_pun" />
+            android:title="@string/pun_jokes" />
         <item
             android:id="@+id/nav_christmas"
             android:icon="@drawable/ic_christmas"
-            android:title="@string/menu_christmas" />
+            android:title="@string/christmas_jokes" />
         <item
             android:id="@+id/nav_spooky"
             android:icon="@drawable/ic_spooky"
-            android:title="@string/menu_spooky" />
+            android:title="@string/spooky_jokes" />
         <item
             android:id="@+id/nav_miscellaneous"
             android:icon="@drawable/ic_bowling"
-            android:title="@string/menu_miscellaneous" />
+            android:title="@string/miscellaneous_jokes" />
     </group>
 
-    <item android:title="@string/text_communicate">
+    <item android:title="@string/communicate">
         <menu>
             <item
                 android:id="@+id/nav_about"
                 android:icon="@drawable/ic_menu_about"
-                android:title="@string/menu_about"
+                android:title="@string/about"
                 android:visible="false"/>
             <item
                 android:id="@+id/nav_settings"
                 android:icon="@drawable/ic_menu_settings"
-                android:title="@string/menu_settings" />
+                android:title="@string/settings" />
         </menu>
     </item>
