+++ app/src/main/res/values-cs/strings.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ Loco xml export: Android string resources
+ Project: lol-project
+ Release: Working copy
+ Locale: cs, Czech
+ Exported by: Segun Francis
+ Exported at: Wed, 06 Oct 2021 15:44:21 +0100
+-->
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+    <string name="any_jokes">Jakékoli vtipy</string>
+    <string name="programming_jokes">Programovací vtipy</string>
+    <string name="dark_jokes">Temné vtipy</string>
+    <string name="miscellaneous_jokes">Různé vtipy</string>
+    <string name="about">O</string>
+    <string name="settings">Nastavení</string>
+    <string name="shuffle">Zamíchat</string>
+    <string name="retry">OPAKUJTE</string>
+    <string name="christmas_jokes">Vánoční vtipy</string>
+    <string name="pun_jokes">Pun vtipy</string>
+    <string name="communicate">Komunikovat</string>
+    <string name="spooky_jokes">Strašidelné vtipy</string>
+    <string name="app_name">LOL</string>
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="nav_header_title">LOL</string>
+    <string name="nav_header_desc">Navigation header</string>
+    <string name="ad_unit_id">ca-app-pub-3927632340219551/3674063347</string>
+    <string name="ad_value">ca-app-pub-3927632340219551~8409700494</string>
+    <string name="are_you_sure_you_want_to_exit">Opravdu chcete odejít?</string>
+    <string name="yes">ANO</string>
+    <string name="no">NE</string>
+</resources>
