+++ app/src/main/res/values-es/strings.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ Loco xml export: Android string resources
+ Project: lol-project
+ Release: Working copy
+ Locale: es, Spanish
+ Exported by: Segun Francis
+ Exported at: Wed, 06 Oct 2021 15:44:21 +0100
+-->
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+    <string name="any_jokes">Cualquier broma</string>
+    <string name="programming_jokes">Programación de chistes</string>
+    <string name="dark_jokes">Chistes oscuros</string>
+    <string name="miscellaneous_jokes">Chistes Misceláneos</string>
+    <string name="about">Sobre</string>
+    <string name="settings">Ajustes</string>
+    <string name="shuffle">Barajar</string>
+    <string name="retry">REVER</string>
+    <string name="christmas_jokes">Chistes de navidad</string>
+    <string name="pun_jokes">Chistes de Retruécano</string>
+    <string name="communicate">Comunicar</string>
+    <string name="spooky_jokes">Chistes espeluznantes</string>
+    <string name="app_name">LOL</string>
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="nav_header_title">LOL</string>
+    <string name="nav_header_desc">Navigation header</string>
+    <string name="ad_unit_id">ca-app-pub-3927632340219551/3674063347</string>
+    <string name="ad_value">ca-app-pub-3927632340219551~8409700494</string>
+    <string name="are_you_sure_you_want_to_exit">¿Estás seguro de que quieres salir?</string>
+    <string name="yes">SÍ</string>
+    <string name="no">NO</string>
+</resources>
