+++ app/src/main/res/values-fr/strings.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ Loco xml export: Android string resources
+ Project: lol-project
+ Release: Working copy
+ Locale: fr, French
+ Exported by: Segun Francis
+ Exported at: Wed, 06 Oct 2021 15:44:21 +0100
+-->
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+    <string name="any_jokes">Toutes les blagues</string>
+    <string name="programming_jokes">Blagues de programmation</string>
+    <string name="dark_jokes">Blagues Sombres</string>
+    <string name="miscellaneous_jokes">Blagues diverses</string>
+    <string name="about">Sur</string>
+    <string name="settings">Paramètres</string>
+    <string name="shuffle">Mélanger</string>
+    <string name="retry">RÉESSAYEZ</string>
+    <string name="christmas_jokes">Blagues de Noël</string>
+    <string name="pun_jokes">Blagues de Calembour</string>
+    <string name="communicate">Communiquer</string>
+    <string name="spooky_jokes">Blagues effrayantes</string>
+    <string name="app_name">LOL</string>
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="nav_header_title">LOL</string>
+    <string name="nav_header_desc">Navigation header</string>
+    <string name="ad_unit_id">ca-app-pub-3927632340219551/3674063347</string>
+    <string name="ad_value">ca-app-pub-3927632340219551~8409700494</string>
+    <string name="are_you_sure_you_want_to_exit">Êtes-vous sûr de vouloir quitter?</string>
+    <string name="yes">OUI</string>
+    <string name="no">NON</string>
+</resources>
