+++ app/src/main/res/values/strings.xml
@@ -1,22 +1,33 @@
-<resources>
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ Loco xml export: Android string resources
+ Project: lol-project
+ Release: Working copy
+ Locale: en-GB, English (Nigeria)
+ Exported by: Segun Francis
+ Exported at: Wed, 06 Oct 2021 15:44:21 +0100
+-->
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+    <string name="any_jokes">Any Jokes</string>
+    <string name="programming_jokes">Programming Jokes</string>
+    <string name="dark_jokes">Dark Jokes</string>
+    <string name="miscellaneous_jokes">Miscellaneous Jokes</string>
+    <string name="about">About</string>
+    <string name="settings">Settings</string>
+    <string name="shuffle">Shuffle</string>
+    <string name="retry">RETRY</string>
+    <string name="christmas_jokes">Christmas Jokes</string>
+    <string name="pun_jokes">Pun Jokes</string>
+    <string name="communicate">Communicate</string>
+    <string name="spooky_jokes">Spooky Jokes</string>
     <string name="app_name">LOL</string>
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
     <string name="nav_header_title">LOL</string>
-    <string name="nav_header_subtitle">android.studio@android.com</string>
     <string name="nav_header_desc">Navigation header</string>
-    <string name="action_settings">Settings</string>
-
-    <string name="menu_any">Any Jokes</string>
-    <string name="menu_programming">Programming Jokes</string>
-    <string name="menu_dark">Dark Jokes</string>
-    <string name="menu_miscellaneous">Miscellaneous Jokes</string>
-    <string name="menu_about">About</string>
-    <string name="menu_settings">Settings</string>
-    <string name="text_shuffle">Shuffle</string>
-    <string name="text_retry">RETRY</string>
-    <string name="menu_spooky">Spooky Jokes</string>
-    <string name="menu_christmas">Christmas Jokes</string>
-    <string name="menu_pun">Pun Jokes</string>
-    <string name="text_communicate">Communicate</string>
+    <string name="ad_unit_id">ca-app-pub-3927632340219551/3674063347</string>
+    <string name="ad_value">ca-app-pub-3927632340219551~8409700494</string>
+    <string name="are_you_sure_you_want_to_exit">Are you sure you want to exit?</string>
+    <string name="yes">YES</string>
+    <string name="no">NO</string>
 </resources>
