+++ app/src/main/res/values/styles.xml
@@ -18,7 +18,7 @@
     <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 
     <style name="FABStyle" parent="Widget.MaterialComponents.ExtendedFloatingActionButton">
-        <item name="android:text">@string/text_shuffle</item>
+        <item name="android:text">@string/shuffle</item>
         <item name="android:textColor">@android:color/white</item>
         <item name="android:textSize">20sp</item>
         <item name="icon">@drawable/ic_shuffle</item>
