+++ build.gradle
@@ -5,15 +5,15 @@ buildscript {
     repositories {
         google()
         mavenCentral()
-        maven {
+        /*maven {
             url 'https://maven.fabric.io/public'
-        }
+        }*/
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.3'
+        classpath "com.android.tools.build:gradle:4.1.3"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.0-M1"
-        classpath 'com.google.gms:google-services:4.3.8'
-        //classpath 'io.fabric.tools:gradle:1.31.2' /* Crashlytics */
+        classpath "com.google.gms:google-services:4.3.10"
+        classpath "com.google.firebase:firebase-crashlytics-gradle:2.7.1"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
