--- secret.properties
@@ -1,1 +1,0 @@
-adUnitId="ca-app-pub-1982812619462025/5967139175"
